+++ com.unity.ml-agents.extensions/Runtime/Input/Adaptors/Vector2InputActionAdaptor.cs
@@ -1,4 +1,4 @@
-#if MLA_INPUT_SYSTEM && UNITY_2019_4_OR_NEWER
+#if MLA_INPUT_SYSTEM
 using System;
 using Unity.MLAgents.Actuators;
 using UnityEngine;
@@ -42,4 +42,4 @@ namespace Unity.MLAgents.Extensions.Input
 
     }
 }
-#endif // MLA_INPUT_SYSTEM && UNITY_2019_4_OR_NEWER
+#endif // MLA_INPUT_SYSTEM