+++ com.unity.ml-agents/Tests/Editor/AcademyTests.cs
@@ -1,10 +1,8 @@
 using NUnit.Framework;
 using Unity.MLAgents.Sensors;
 using UnityEngine;
-#if UNITY_2019_3_OR_NEWER
 using System.Reflection;
 using Unity.MLAgents;
-#endif
 
 namespace Unity.MLAgents.Tests
 {
@@ -14,13 +12,9 @@ namespace Unity.MLAgents.Tests
         [Test]
         public void TestPackageVersion()
         {
-            // Make sure that the version strings in the package and Academy don't get out of sync.
-            // Unfortunately, the PackageInfo methods don't exist in earlier versions of the editor.
-#if UNITY_2019_3_OR_NEWER
             var packageInfo = UnityEditor.PackageManager.PackageInfo.FindForAssembly(typeof(Agent).Assembly);
             Assert.AreEqual("com.unity.ml-agents", packageInfo.name);
             Assert.AreEqual(Academy.k_PackageVersion, packageInfo.version);
-#endif
         }
 
         class RecursiveAgent : Agent